pyproject.toml
[tool.poetry]
name = "mcqq-tool"
version = "1.2.6"
description = "MC_QQ工具包"
authors = ["17TheWord <17theword@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.3"
nonebot-adapter-qq = "^1.5.1"
nonebot-adapter-onebot = "^2.4.4"
nonebot-adapter-minecraft = "^1.3.1"
nonebot-plugin-guild-patch-remix = "^0.2.4"

[tool.poetry.group.test.dependencies]
ruff = "^0.4.5"
isort = "^5.10.1"
nonebug = "^0.3.5"
pytest-cov = "^4.0.0"
pre-commit = "^3.7.0"
pytest-asyncio = "^0.23.0"
nonebot2 = { version = "^2.2.0", extras = ["fastapi"] }

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "PYI", "PT", "Q"]
ignore = ["E402", "E501", "F401", "C901", "UP037", "UP035", "UP006", "PYI021"]

[tool.ruff.lint.per-file-ignores]
"mcqq_tool/__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.data.packages]
my_package = { path = "mcqq_tool" }
